+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## 2.1.0
 
+ADDED:
+  - Ajout de la route GET /admin/1.0.0/services dans l'API d'administration
+
 CHANGED:
   - La documentation de l'API d'administration a été grandement enrichie. 
   - La route /health a une réponse plus complète et est vraiment codée pour prendre en compte l'état de chaque service et chaque source disponibles.